Error messages (#551)
* Improve responses for chains and assets
* Add tests for invalid requests.
* Replace content type for encryptashex.
* Reorganise tests.
* First batch for improving test coverage.
* Whole coverage to assets.
* Full coverage to chains file.
