+++ aquarius/app/chains.py
@@ -6,7 +6,7 @@ import elasticsearch
 import json
 import logging
 
-from flask import Blueprint, Response
+from flask import Blueprint, jsonify
 from aquarius.app.es_instance import ElasticsearchInstance
 from aquarius.log import setup_logging
 from aquarius.myapp import app
@@ -33,13 +33,13 @@ def get_chains_list():
         chains = es_instance.es.get(
             index=f"{es_instance.db_index}_plus", id="chains", doc_type="_doc"
         )["_source"]
-        return Response(json.dumps(chains), 200, content_type="application/json")
+        return json.dumps(chains)
     except (elasticsearch.exceptions.NotFoundError, KeyError):
-        logger.error("Cannot get chains list")
-        return Response("No chains found", 404)
+        logger.error("Cannot get chains list.")
+        return jsonify(error="No chains found."), 404
     except Exception as e:
-        logger.error(f"Cannot get chains list: {str(e)}")
-        return Response("No chains found", 404)
+        logger.error(f"Error in get_chains_list: {str(e)}")
+        return jsonify(error=f"Error retrieving chains: {str(e)}."), 404
 
 
 @chains.route("/status/<chain_id>", methods=["GET"])
@@ -66,12 +66,12 @@ def get_index_status(chain_id):
             id="events_last_block_" + str(chain_id),
             doc_type="_doc",
         )["_source"]
-        return Response(
-            json.dumps(last_block_record), 200, content_type="application/json"
-        )
+        return json.dumps(last_block_record)
     except (elasticsearch.exceptions.NotFoundError, KeyError):
-        logger.error(f"Cannot get index status for chain {chain_id}")
-        return Response(f"{chain_id} is not indexed", 404)
+        logger.error(f"Cannot get index status for chain {chain_id}. Chain not found.")
+        return jsonify(error=f"Chain {chain_id} is not indexed."), 404
     except Exception as e:
-        logger.error(f"Cannot get index status for chain {chain_id}: {str(e)}")
-        return Response(f"{chain_id} is not indexed", 404)
+        logger.error(
+            f"Cannot get index status for chain {chain_id}. Error encountered is: {str(e)}"
+        )
+        return jsonify(error=f"Error retrieving chain {chain_id}: {str(e)}."), 404
