+++ tests/test_util.py
@@ -11,6 +11,7 @@ from aquarius.app.util import (
     get_bool_env_value,
     datetime_converter,
     check_no_urls_in_files,
+    validate_date_format,
 )
 from aquarius.app.auth_util import compare_eth_addresses
 from aquarius.block_utils import BlockProcessingClass
@@ -87,3 +88,16 @@ def test_check_no_urls_in_files_fails():
     message, code = check_no_urls_in_files(main, "GET")
     assert message == "GET request failed: url is not allowed in files "
     assert code == 400
+
+
+def test_date_format_validator():
+    date = "2016-02-08T16:02:20Z"
+    assert validate_date_format(date) == (None, None)
+
+
+def test_invalid_date():
+    date = "XXXX"
+    assert validate_date_format(date) == (
+        "Incorrect data format, should be '%Y-%m-%dT%H:%M:%SZ'",
+        400,
+    )
